backend/requirements.txt
anyio==3.3.4
asgiref==3.4.1
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
ecdsa==0.17.0
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyMySQL==1.0.2
python-jose==3.3.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
typing-extensions==3.10.0.2
uvicorn==0.15.0
